+++ tests/tasks/test_create_project.py
@@ -7,7 +7,7 @@ from airflow.utils.state import DagRunState
 from airflow.utils.types import DagRunType
 
 from toloka.client import Project, structure
-import toloka_provider.tasks.toloka as tlk_ops
+import toloka_provider.tasks.toloka as tlk_tasks
 
 from ..time_config import DATA_INTERVAL_START, DATA_INTERVAL_END
 
@@ -101,7 +101,7 @@ def dag_for_test_project_creation(raw_project_map, simple_project_map):
     @dag(schedule_interval='@once', default_args={'start_date': DATA_INTERVAL_START})
     def dag_project():
         project = prepare_project()
-        created_project = tlk_ops.create_project(obj=project, toloka_conn_id='toloka_conn')
+        created_project = tlk_tasks.create_project(obj=project, toloka_conn_id='toloka_conn')
         check_project(created_project)
 
     return dag_project()