+++ CHANGELOG.md
@@ -1,7 +1,16 @@
+0.0.4
+-------------------
+* Rename package to `airflow-provider-toloka`
+* Structure package by standard
+* Rename operators to funcops
+* Remove `get_assignments_df`
+
+
 0.0.3
 -------------------
 * Fixed `open_pool` issues
 * Added tests with code coverage
+
 
 0.0.2 [YANKED]
 -------------------